Make port arg optional in `docker compose port`
This comand no longer needs a mandatory port. If no port is provided it
will instead list all published ports. This behavior falls in line with
the behavior provided by `docker port`.

Closes #11859.

+++ cmd/formatter/container.go
@@ -216,8 +216,8 @@ func (c *ContainerContext) Ports() string {
 	for _, publisher := range c.c.Publishers {
 		ports = append(ports, types.Port{
 			IP:          publisher.URL,
-			PrivatePort: uint16(publisher.TargetPort),
-			PublicPort:  uint16(publisher.PublishedPort),
+			PrivatePort: publisher.TargetPort,
+			PublicPort:  publisher.PublishedPort,
 			Type:        publisher.Protocol,
 		})
 	}
